Changed small hotfixes

+++ Shuttle-back/src/main/java/com/shuttle/location/dto/LocationDTO.java
@@ -2,8 +2,8 @@ package com.shuttle.location.dto;
 
 public class LocationDTO {
 	private String address;
-	private long latitude;
-	private long longitude;
+	private double latitude;
+	private double longitude;
 	
 	public String getAddress() {
 		return address;
@@ -11,17 +11,18 @@ public class LocationDTO {
 	public void setAddress(String adress) {
 		this.address = adress;
 	}
-	public long getLatitude() {
+	public double getLatitude() {
 		return latitude;
 	}
-	public void setLatitude(long latitude) {
+	public void setLatitude(double latitude) {
 		this.latitude = latitude;
 	}
-	public long getLongitude() {
+	public double getLongitude() {
 		return longitude;
 	}
-	public void setLongitude(long longitude) {
+	public void setLongitude(double longitude) {
 		this.longitude = longitude;
 	}
+	
 	
 }
